server/requirements.txt
django==4.2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-api-logger==1.1.12
drf-spectacular==0.26.2
django-debug-toolbar==4.1.0
psycopg2-binary==2.9.6
requests==2.31.0
gunicorn==20.1.0
loguru==0.7.0
python-decouple==3.8
